# apps/agent-server/openapi.yaml
openapi: 3.1.0
info:
  title: Robota Agent Server API
  version: 1.0.0
  description: |
    Remote AI provider proxy server with Playground WebSocket support.
    Deployable standalone or as Firebase Functions.

    WebSocket endpoint `WS /ws/playground` is not described here; see the
    `PlaygroundWebSocketServer` class contract in `docs/SPEC.md` for the
    auth/message protocol.

servers:
  - url: http://localhost:{port}
    description: Local development server
    variables:
      port:
        default: '3000'
        description: Server port (set via PORT environment variable)

paths:
  /:
    get:
      operationId: getServerInfo
      summary: Server info
      description: Returns server name, version, available endpoints, and registered providers.
      responses:
        '200':
          description: Server info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerInfo'

  /health:
    get:
      operationId: getHealth
      summary: Global health check
      description: Returns server liveness status and environment metadata.
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /api/v1/remote/health:
    get:
      operationId: getRemoteHealth
      summary: Remote provider health check
      description: Returns OK status and the list of registered AI provider names.
      responses:
        '200':
          description: Providers available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoteHealthResponse'

  /api/v1/remote/chat:
    post:
      operationId: postRemoteChat
      summary: Provider chat proxy
      description: |
        Proxy a chat-completion request to a registered AI provider.
        The provider must be registered server-side via a corresponding `*_API_KEY` environment
        variable. Returns the raw provider response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Provider response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoteChatResponse'
        '400':
          description: Invalid request (missing provider, unknown provider, or invalid messages)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'
        '500':
          description: Provider or internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/remote/chat/stream:
    post:
      operationId: postRemoteChatStream
      summary: Streaming provider chat proxy
      description: |
        Proxy a chat request over SSE. The server emits `delta` frames, one assembled `message`
        frame, then for a selected model effort exactly one server-adapter-owned
        `model-effort-outcome` frame before `done`. The client keeps its callback local and only
        transports that terminal value.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: SSE frames (`delta`, `message`, `model-effort-outcome`, `done`, or `error`)
          content:
            text/event-stream:
              schema:
                type: string
                description: UTF-8 Server-Sent Events; each `data:` value uses the named event schema.
              x-sse-events:
                delta:
                  $ref: '#/components/schemas/RemoteChatStreamDelta'
                message:
                  $ref: '#/components/schemas/AssistantMessage'
                model-effort-outcome:
                  $ref: '#/components/schemas/ModelEffortOutcome'
                done:
                  $ref: '#/components/schemas/RemoteChatStreamDone'
                error:
                  $ref: '#/components/schemas/RemoteChatStreamError'
        '400':
          description: Invalid request before the SSE stream is opened
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Provider or internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/remote/ws/status:
    get:
      operationId: getWsStatus
      summary: WebSocket server status
      description: Returns connection statistics for the Playground WebSocket server.
      responses:
        '200':
          description: WebSocket status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WsStatusResponse'

components:
  schemas:
    ServerInfo:
      type: object
      required: [name, version, description, environment, timestamp, endpoints, providers]
      properties:
        name:
          type: string
          example: Robota SDK API Server
        version:
          type: string
          example: '1.0.0'
        description:
          type: string
        environment:
          type: string
          example: development
        timestamp:
          type: string
          format: date-time
        endpoints:
          type: object
          properties:
            health:
              type: string
              example: /api/v1/remote/health
            chat:
              type: string
              example: /api/v1/remote/chat
        providers:
          type: array
          items:
            type: string
          example: [openai, anthropic]

    HealthResponse:
      type: object
      required: [status, timestamp, service, version, environment]
      properties:
        status:
          type: string
          enum: [ok]
        timestamp:
          type: string
          format: date-time
        service:
          type: string
          example: robota-agent-server
        version:
          type: string
          example: '1.0.0'
        environment:
          type: string
          example: production

    RemoteHealthResponse:
      type: object
      required: [status, providers, timestamp]
      properties:
        status:
          type: string
          enum: [ok]
        providers:
          type: array
          items:
            type: string
          example: [openai, anthropic, google]
        timestamp:
          type: string
          format: date-time

    UniversalMessage:
      type: object
      required: [role]
      description: Universal message format (agent-core TUniversalMessage).
      properties:
        role:
          type: string
          enum: [user, assistant, system, tool]
        content:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
        tool_calls:
          type: array
          items:
            type: object
        tool_call_id:
          type: string

    ChatRequest:
      type: object
      required: [provider, messages]
      properties:
        provider:
          type: string
          description: Registered provider name (e.g. "openai", "anthropic", "google")
          example: openai
        messages:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/UniversalMessage'
          description: Non-empty message array in universal message format
        model:
          type: string
          description: Optional model override. Defaults to provider package default.
          example: gpt-4o
        tools:
          type: array
          items:
            type: object
            additionalProperties: true
        options:
          $ref: '#/components/schemas/ChatOptions'

    ChatOptions:
      type: object
      properties:
        maxTokens:
          type: number
        temperature:
          type: number
        effort:
          $ref: '#/components/schemas/ModelEffortSelection'
        toolChoice:
          oneOf:
            - type: string
              enum: [auto, none, required]
            - type: object
              required: [tool]
              properties:
                tool:
                  type: string
        responseFormat:
          type: object
          additionalProperties: true
        nativeWebTools:
          type: object
          additionalProperties: true
        openai:
          type: object
          additionalProperties: true
        anthropic:
          type: object
          additionalProperties: true
        google:
          type: object
          additionalProperties: true

    ModelEffortSelection:
      type: string
      enum: [auto, none, minimal, low, medium, high, xhigh, max]

    ModelEffortOutcome:
      type: object
      required: [resolution, nativeControl, providerDispatch]
      properties:
        resolution:
          type: object
          required: [selection, effective, disposition, fingerprint]
          properties:
            selection:
              $ref: '#/components/schemas/ModelEffortSelection'
            effective:
              oneOf:
                - type: 'null'
                - type: string
                  enum: [none, minimal, low, medium, high, xhigh, max]
            disposition:
              type: string
              enum: [exact, clamped, model-default, not-applied]
            fingerprint:
              type: string
        nativeControl:
          oneOf:
            - type: object
              required: [state, id]
              properties:
                state:
                  type: string
                  enum: [sent]
                id:
                  type: string
            - type: object
              required: [state, reason]
              properties:
                state:
                  type: string
                  enum: [omitted]
                reason:
                  type: string
        providerDispatch:
          oneOf:
            - type: object
              required: [state]
              properties:
                state:
                  type: string
                  enum: [sent]
            - type: object
              required: [state, reason]
              properties:
                state:
                  type: string
                  enum: [not-dispatched]
                reason:
                  type: string

    TokenUsage:
      type: object
      required: [promptTokens, completionTokens, totalTokens]
      properties:
        promptTokens:
          type: integer
        completionTokens:
          type: integer
        totalTokens:
          type: integer

    AssistantMessage:
      type: object
      required: [id, role, content, state, timestamp]
      description: Serialized agent-core assistant message returned by a provider chat call.
      properties:
        id:
          type: string
        role:
          type: string
          enum: [assistant]
        content:
          type: string
          nullable: true
          description: Generated text content, or null for tool-call-only responses.
        toolCalls:
          type: array
          items:
            type: object
          description: Tool calls requested by the model.
        state:
          type: string
          enum: [complete, interrupted]
        timestamp:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true

    RemoteChatResponse:
      allOf:
        - $ref: '#/components/schemas/AssistantMessage'
        - type: object
          properties:
            modelEffortOutcome:
              $ref: '#/components/schemas/ModelEffortOutcome'

    RemoteChatStreamDelta:
      type: object
      required: [text]
      properties:
        text:
          type: string

    RemoteChatStreamDone:
      type: string
      enum: ['[DONE]']

    RemoteChatStreamError:
      type: object
      required: [error]
      properties:
        error:
          type: string

    WsStatusEnabled:
      type: object
      required: [websocket]
      properties:
        websocket:
          type: object
          required:
            [
              enabled,
              endpoint,
              totalConnections,
              authenticatedConnections,
              uniqueUsers,
              uniqueSessions,
            ]
          properties:
            enabled:
              type: boolean
              enum: [true]
            endpoint:
              type: string
              example: /ws/playground
            totalConnections:
              type: integer
            authenticatedConnections:
              type: integer
            uniqueUsers:
              type: integer
            uniqueSessions:
              type: integer

    WsStatusDisabled:
      type: object
      required: [websocket]
      properties:
        websocket:
          type: object
          required: [enabled, message]
          properties:
            enabled:
              type: boolean
              enum: [false]
            message:
              type: string
              example: WebSocket server not initialized

    WsStatusResponse:
      oneOf:
        - $ref: '#/components/schemas/WsStatusEnabled'
        - $ref: '#/components/schemas/WsStatusDisabled'

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          oneOf:
            - type: string
              description: Simple error message string (400 responses)
            - type: object
              required: [message, status, timestamp]
              properties:
                message:
                  type: string
                status:
                  type: integer
                timestamp:
                  type: string
                  format: date-time

    RateLimitError:
      type: object
      required: [error, retryAfter]
      properties:
        error:
          type: string
          example: Too many requests
        retryAfter:
          type: string
          example: 15 minutes